Revert cl/281593369.
After cl/292973383, we can no longer rely on the import resolver to dynamically add files to the project as they are resolved. So we have to go back to attaching to every proto library transitively for go proto sources. Will increase initial indexing time as more .pb.go files will be initially present in the external library, whether they are actually used or not.

PiperOrigin-RevId: 305729013